Fixes https://github.com/nethesis/dev/issues/6188

+++ NethLED/Form1.vb
@@ -6,6 +6,7 @@ Public Class Form1
     Private WithEvents Watch As System.IO.FileSystemWatcher
     Private Device As IDevice
 
+
     Private Sub Form1_Load(sender As Object, e As EventArgs) Handles MyBase.Load
         ExceptionManager = New ExceptionManager
         Me.Opacity = 0
@@ -37,6 +38,16 @@ Public Class Form1
 
         With GetLED()
             .Blink(LedTarget.All, New Color(255, 69, 0), 5, 0, 0)
+            Threading.Thread.Sleep(1000)
+            .SetColor(LedTarget.All, New Color(0, 0, 0))
+            Threading.Thread.Sleep(1000)
+        End With
+
+    End Sub
+
+    Private Sub OtherBlinking(RGB As String())
+        With GetLED()
+            .Blink(LedTarget.All, New Color(CInt(RGB(0)), CInt(RGB(1)), CInt(RGB(2))), 5, 0, 0)
             Threading.Thread.Sleep(1000)
             .SetColor(LedTarget.All, New Color(0, 0, 0))
             Threading.Thread.Sleep(1000)
@@ -76,41 +87,48 @@ Public Class Form1
 
     Private Sub LED_Tick(sender As Object, e As EventArgs) Handles LED.Tick
         LED.Stop()
+        If LED_Type <> "" Then
+            Try
+                Dim F As String = IO.Path.Combine(Path, "led_" & LED_Type)
+                Debug.WriteLine("NethLed Type: " & LED_Type)
+                If IO.File.Exists(F) Then
+                    IO.File.Delete(F)
+                End If
 
-        Try
-            Dim F As String = IO.Path.Combine(Path, "led_" & LED_Type)
-            If IO.File.Exists(F) Then
-                IO.File.Delete(F)
-            End If
+                Select Case LED_Type
+                    Case "incoming"
+                        RingBlinking()
+                    Case "busy"
+                        Busy()
+                        'LED_Type = ""
+                    Case "off", "offline"
+                        SwitchOffLED()
+                        'LED_Type = ""
+                    Case "online"
+                        Online()
+                        'LED_Type = ""
+                    Case "available"
+                        Available()
+                        'LED_Type = ""
 
-            Select Case LED_Type
-                Case "incoming"
-                    RingBlinking()
-                Case "busy"
-                    Busy()
-                    LED_Type = ""
-                Case "off", "offline"
-                    SwitchOffLED()
-                    LED_Type = ""
-                Case "online"
-                    Online()
-                    LED_Type = ""
-                Case "available"
-                    Available()
-                    LED_Type = ""
-
-                Case Else
-                    Dim RGB As String() = Split(LED_Type, "_")
-                    With GetLED()
-                        Threading.Thread.Sleep(1)
-                        '.SetColor(LedTarget.All, New Color(CInt(RGB(0)), CInt(RGB(1)), CInt(RGB(2))))
-                        .SetColor(LedTarget.All, New Color(0, 1, 2))
-                    End With
-                    LED_Type = ""
-            End Select
-        Catch ex As Exception
-
-        End Try
+                    Case Else
+                        Dim RGB As String() = Split(LED_Type, "_")
+                        With GetLED()
+                            Threading.Thread.Sleep(1)
+                            If RGB.Contains("blink") Then
+                                .Blink(LedTarget.All, New Color(CInt(RGB(0)), CInt(RGB(1)), CInt(RGB(2))), 30, 0, 0)
+                            Else
+                                .SetColor(LedTarget.All, New Color(CInt(RGB(0)), CInt(RGB(1)), CInt(RGB(2))))
+                            End If
+                            '.SetColor(LedTarget.All, New Color(0, 1, 2))
+                        End With
+                        'OtherBlinking(RGB)
+                        'LED_Type = ""
+                End Select
+            Catch ex As Exception
+                Debug.WriteLine("NethLed:" & LED_Type)
+            End Try
+        End If
 
         LED.Start()
     End Sub
@@ -118,16 +136,20 @@ Public Class Form1
 
     Private Sub OnError(sender As Object, e As ErrorEventArgs)
         WaitAndRecover(e)
+        Debug.WriteLine("NethLed: OnError")
     End Sub
 
     Private Path As String = ""
     Private LED_Type As String = ""
     Private Sub OnCreated(sender As Object, e As FileSystemEventArgs)
+        Debug.WriteLine("NethLed: OnCreated " & e.Name.ToLower)
         If e.Name.ToLower.StartsWith("led_") Then
             LED_Type = e.Name.Substring(4)
+            Debug.WriteLine("LED_Type :" & LED_Type)
         End If
     End Sub
     Private Sub OnDeleted(sender As Object, e As FileSystemEventArgs)
+        Debug.WriteLine("NethLed: OnDeleted")
         'If e.Name.ToLower.StartsWith("led_" & LED_Type) Then
         '    LED_Type = "off"
         'End If
@@ -139,6 +161,7 @@ Public Class Form1
         If My.Application.CommandLineArgs.Contains("-e") Then
             Path = IO.Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), Application.ProductName)
         End If
+        Debug.WriteLine("NethLed watcher on: " & Path)
 
         Watch = New FileSystemWatcher
         With Watch
@@ -167,6 +190,7 @@ Public Class Form1
         Watch = New FileSystemWatcher
         While Not Watch.EnableRaisingEvents
             Try
+                Debug.WriteLine("NethLed: Error ")
                 StartWatching()
 
                 If Not Watch.EnableRaisingEvents Then
